// libs/plugins/ddd/src/generators/feature/schema.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface FeatureOptions {
  /**
   * Feature name
   */
  name: string;
  /**
   * Domain name
   */
  domain?: string;
  /**
   * LeanIX name
   */
  leanIxName: string;
  /**
   * MFE name
   */
  mfeName: string;
  /**
   * Add ngrx for the domain (entity required)
   */
  ngrx?: boolean;
  [k: string]: unknown;
}